requirements.txt
djangorestframework>3.4.0
rethinkdb
deepdiff>3.2.0
